@@ -1,16 +1,3 @@
 module github.com/knwoop/dynamic-struct
 
-go 1.17
-
-require (
-	github.com/goccy/go-yaml v1.9.4
-	github.com/google/go-cmp v0.5.6
-)
-
-require (
-	github.com/fatih/color v1.10.0 // indirect
-	github.com/mattn/go-colorable v0.1.8 // indirect
-	github.com/mattn/go-isatty v0.0.12 // indirect
-	golang.org/x/sys v0.0.0-20200223170610-d5e6a3e2c0ae // indirect
-	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
-)
+go 1.24
