@@ -1,34 +1,48 @@
 package dynamicstruct
 
 import (
+	"errors"
+	"fmt"
+	"go/token"
 	"reflect"
+	"sort"
 )
 
-// Builder builds struct dynamically
+// Builder builds a struct dynamically.
 type Builder struct {
 	fields map[string]*Field
+	errs   []error
 }
 
-// NewBuilder returns new clean instance of Builder interface
-// builder := dynamicstruct.NewStruct()
+// NewBuilder returns a new clean instance of Builder.
+// builder := dynamicstruct.NewBuilder()
 func NewBuilder() *Builder {
 	return &Builder{
 		fields: map[string]*Field{},
 	}
 }
 
-// AddField adds a field to a struct
-func (b *Builder) AddField(name string, typ interface{}, tag string) *Builder {
-	b.fields[name] = &Field{
-		Name: name,
-		Type: typ,
-		Tag:  tag,
+// AddField adds a field to the struct being built.
+// It validates the field name and type, accumulating errors
+// that are reported when Build is called.
+func (b *Builder) AddField(name string, typ any, tag string) *Builder {
+	if name == "" {
+		b.errs = append(b.errs, errors.New("field name must not be empty"))
+		return b
 	}
-
+	if !token.IsExported(name) {
+		b.errs = append(b.errs, fmt.Errorf("field name %q must be exported (start with uppercase)", name))
+		return b
+	}
+	if typ == nil {
+		b.errs = append(b.errs, fmt.Errorf("field %q: type must not be nil", name))
+		return b
+	}
+	b.fields[name] = &Field{Name: name, Type: typ, Tag: tag}
 	return b
 }
 
-// RemoveField deletes a field held in Builder
+// RemoveField deletes a field held in Builder.
 func (b *Builder) RemoveField(name string) *Builder {
 	delete(b.fields, name)
 	return b
@@ -49,8 +63,13 @@ func (b *Builder) GetField(name string) *Field {
 	return b.fields[name]
 }
 
-// Build returns a dynamically created struct
-func (b *Builder) Build() *DynamicStruct {
+// Build returns a dynamically created struct definition.
+// It returns an error if any validation errors were accumulated during AddField calls.
+func (b *Builder) Build() (*DynamicStruct, error) {
+	if len(b.errs) > 0 {
+		return nil, errors.Join(b.errs...)
+	}
+
 	structFields := make([]reflect.StructField, 0, len(b.fields))
 	for name, field := range b.fields {
 		structFields = append(structFields, reflect.StructField{
@@ -60,17 +79,22 @@ func (b *Builder) Build() *DynamicStruct {
 		})
 	}
 
+	// Sort fields by name for deterministic ordering.
+	sort.Slice(structFields, func(i, j int) bool {
+		return structFields[i].Name < structFields[j].Name
+	})
+
 	return &DynamicStruct{
 		definition: reflect.StructOf(structFields),
-	}
+	}, nil
 }
 
-// DynamicStruct represents a struct created dynamically by Builder
+// DynamicStruct represents a struct created dynamically by Builder.
 type DynamicStruct struct {
 	definition reflect.Type
 }
 
-// New creates dynamic struct
-func (ds *DynamicStruct) New() interface{} {
+// New creates a new instance of the dynamic struct.
+func (ds *DynamicStruct) New() any {
 	return reflect.New(ds.definition).Interface()
 }
