@@ -1,63 +1,8 @@
 package dynamicstruct
 
-// A FieldType indicates which member of the Field union struct should be used
-// and how it should be serialized.
-type FieldType uint8
-
-const (
-	// UnknownType is the default field type.
-	UnknownType FieldType = iota
-	// BinaryType indicates that the field carries an opaque binary blob.
-	BinaryType
-	// BoolType indicates that the field carries a bool.
-	BoolType
-	// ByteStringType indicates that the field carries UTF-8 encoded bytes.
-	ByteStringType
-	// Complex128Type indicates that the field carries a complex128.
-	Complex128Type
-	// Complex64Type indicates that the field carries a complex128.
-	Complex64Type
-	// DurationType indicates that the field carries a time.Duration.
-	DurationType
-	// Float64Type indicates that the field carries a float64.
-	Float64Type
-	// Float32Type indicates that the field carries a float32.
-	Float32Type
-	// Int64Type indicates that the field carries an int64.
-	Int64Type
-	// Int32Type indicates that the field carries an int32.
-	Int32Type
-	// Int16Type indicates that the field carries an int16.
-	Int16Type
-	// Int8Type indicates that the field carries an int8.
-	Int8Type
-	// StringType indicates that the field carries a string.
-	StringType
-	// TimeType indicates that the field carries a time.Time that is
-	// representable by a UnixNano() stored as an int64.
-	TimeType
-	// TimeFullType indicates that the field carries a time.Time stored as-is.
-	TimeFullType
-	// Uint64Type indicates that the field carries a uint64.
-	Uint64Type
-	// Uint32Type indicates that the field carries a uint32.
-	Uint32Type
-	// Uint16Type indicates that the field carries a uint16.
-	Uint16Type
-	// Uint8Type indicates that the field carries a uint8.
-	Uint8Type
-	// UintptrType indicates that the field carries a uintptr.
-	UintptrType
-	// UintType indicates that the field carries a uintptr.
-	UintType
-	// ErrorType indicates that the field carries an error.
-	ErrorType
-	// SkipType indicates that the field is a no-op.
-	SkipType
-)
-
+// Field represents a single field in a dynamic struct.
 type Field struct {
 	Name string
-	Type interface{}
+	Type any
 	Tag  string
 }
