@@ -1,20 +1,21 @@
 package dynamicstruct
 
 import (
+	"encoding/json"
+	"reflect"
+	"strings"
 	"testing"
-
-	"github.com/google/go-cmp/cmp"
 )
 
 func TestNewBuilder(t *testing.T) {
 	got := NewBuilder()
 
 	if got.fields == nil {
-		t.Error(`TestNewBuilder - expected instance of map[string]*Field got nil`)
+		t.Error("expected instance of map[string]*Field got nil")
 	}
 
 	if len(got.fields) > 0 {
-		t.Errorf(`TestNewGot - expected length of fields map to be 0 got %d`, len(got.fields))
+		t.Errorf("expected length of fields map to be 0 got %d", len(got.fields))
 	}
 }
 
@@ -25,7 +26,7 @@ func TestBuilder_AddField(t *testing.T) {
 
 	got, ok := builder.fields["Field"]
 	if !ok {
-		t.Fatal(`TestBuilder_AddField - expected to have field "Field"`)
+		t.Fatal(`expected to have field "Field"`)
 	}
 
 	expected := &Field{
@@ -34,8 +35,8 @@ func TestBuilder_AddField(t *testing.T) {
 		Tag:  `key:"value"`,
 	}
 
-	if diff := cmp.Diff(got, expected); diff != "" {
-		t.Errorf(`TestBuilder_AddField - (-got +want)\n%v`, diff)
+	if !reflect.DeepEqual(got, expected) {
+		t.Errorf("AddField mismatch: got %+v, want %+v", got, expected)
 	}
 }
 
@@ -45,20 +46,20 @@ func TestBuilder_RemoveField(t *testing.T) {
 	builder.RemoveField("Field")
 
 	if _, ok := builder.fields["Field"]; ok {
-		t.Error(`TestBuilder_RemoveField - expected not to have field "Field"`)
+		t.Error(`expected not to have field "Field"`)
 	}
 }
 
 func TestBuilder_HasField(t *testing.T) {
 	builder := NewBuilder()
 	if builder.HasField("Field") {
-		t.Error(`TestBuilder_HasField - expected not to have field "Field"`)
+		t.Error(`expected not to have field "Field"`)
 	}
 
 	builder.AddField("Field", 1, `key:"value"`)
 
 	if !builder.HasField("Field") {
-		t.Error(`TestBuilder_HasField - expected to have field "Field"`)
+		t.Error(`expected to have field "Field"`)
 	}
 }
 
@@ -66,7 +67,7 @@ func TestBuilder_GetField(t *testing.T) {
 	builder := NewBuilder()
 	field := builder.GetField("Field")
 	if field != nil {
-		t.Errorf(`TestBuilder_GetField - expected instance of map[string]*Field got nil`)
+		t.Error("expected nil for non-existent field")
 	}
 
 	builder.AddField("Field", 1, `key:"value"`)
@@ -76,7 +77,188 @@ func TestBuilder_GetField(t *testing.T) {
 		Tag:  `key:"value"`,
 	}
 	got := builder.GetField("Field")
-	if diff := cmp.Diff(got, expected); diff != "" {
-		t.Errorf(`TestBuilder_AddField - (-got +want)\n%v`, diff)
+	if !reflect.DeepEqual(got, expected) {
+		t.Errorf("GetField mismatch: got %+v, want %+v", got, expected)
+	}
+}
+
+func TestBuilder_Build(t *testing.T) {
+	ds, err := NewBuilder().
+		AddField("Name", "", `json:"name"`).
+		AddField("Age", 0, `json:"age"`).
+		Build()
+	if err != nil {
+		t.Fatalf("Build() unexpected error: %v", err)
+	}
+
+	typ := ds.definition
+
+	if typ.Kind() != reflect.Struct {
+		t.Fatalf("expected struct kind, got %v", typ.Kind())
+	}
+
+	if typ.NumField() != 2 {
+		t.Fatalf("expected 2 fields, got %d", typ.NumField())
+	}
+
+	// Fields should be sorted by name: Age, Name
+	f0 := typ.Field(0)
+	if f0.Name != "Age" {
+		t.Errorf("expected first field to be Age, got %s", f0.Name)
+	}
+	if f0.Tag != `json:"age"` {
+		t.Errorf("expected tag json:\"age\", got %s", f0.Tag)
+	}
+
+	f1 := typ.Field(1)
+	if f1.Name != "Name" {
+		t.Errorf("expected second field to be Name, got %s", f1.Name)
+	}
+}
+
+func TestBuilder_New(t *testing.T) {
+	ds, err := NewBuilder().
+		AddField("Name", "", `json:"name"`).
+		AddField("Age", 0, `json:"age"`).
+		Build()
+	if err != nil {
+		t.Fatalf("Build() unexpected error: %v", err)
+	}
+
+	instance := ds.New()
+
+	// Should be a pointer to a struct
+	rv := reflect.ValueOf(instance)
+	if rv.Kind() != reflect.Ptr || rv.Elem().Kind() != reflect.Struct {
+		t.Fatalf("expected pointer to struct, got %v", rv.Type())
+	}
+
+	elem := rv.Elem()
+
+	// Set values via reflection
+	elem.FieldByName("Name").SetString("Alice")
+	elem.FieldByName("Age").SetInt(30)
+
+	// Read them back
+	if got := elem.FieldByName("Name").String(); got != "Alice" {
+		t.Errorf("expected Name=Alice, got %s", got)
+	}
+	if got := elem.FieldByName("Age").Int(); got != 30 {
+		t.Errorf("expected Age=30, got %d", got)
+	}
+}
+
+func TestBuilder_New_JSON(t *testing.T) {
+	ds, err := NewBuilder().
+		AddField("Name", "", `json:"name"`).
+		AddField("Age", 0, `json:"age"`).
+		Build()
+	if err != nil {
+		t.Fatalf("Build() unexpected error: %v", err)
+	}
+
+	instance := ds.New()
+
+	data := `{"name":"Bob","age":25}`
+	if err := json.Unmarshal([]byte(data), instance); err != nil {
+		t.Fatalf("Unmarshal error: %v", err)
+	}
+
+	elem := reflect.ValueOf(instance).Elem()
+	if got := elem.FieldByName("Name").String(); got != "Bob" {
+		t.Errorf("expected Name=Bob, got %s", got)
+	}
+	if got := elem.FieldByName("Age").Int(); got != 25 {
+		t.Errorf("expected Age=25, got %d", got)
+	}
+
+	// Marshal back
+	out, err := json.Marshal(instance)
+	if err != nil {
+		t.Fatalf("Marshal error: %v", err)
+	}
+	if string(out) != `{"age":25,"name":"Bob"}` {
+		t.Errorf("unexpected JSON output: %s", string(out))
+	}
+}
+
+func TestBuilder_Validation_EmptyName(t *testing.T) {
+	_, err := NewBuilder().
+		AddField("", "", "").
+		Build()
+	if err == nil {
+		t.Fatal("expected error for empty field name")
+	}
+	if !strings.Contains(err.Error(), "must not be empty") {
+		t.Errorf("unexpected error message: %v", err)
+	}
+}
+
+func TestBuilder_Validation_UnexportedName(t *testing.T) {
+	_, err := NewBuilder().
+		AddField("lowercase", "", "").
+		Build()
+	if err == nil {
+		t.Fatal("expected error for unexported field name")
+	}
+	if !strings.Contains(err.Error(), "must be exported") {
+		t.Errorf("unexpected error message: %v", err)
+	}
+}
+
+func TestBuilder_Validation_NilType(t *testing.T) {
+	_, err := NewBuilder().
+		AddField("Field", nil, "").
+		Build()
+	if err == nil {
+		t.Fatal("expected error for nil type")
+	}
+	if !strings.Contains(err.Error(), "must not be nil") {
+		t.Errorf("unexpected error message: %v", err)
+	}
+}
+
+func TestBuilder_Validation_MultipleErrors(t *testing.T) {
+	_, err := NewBuilder().
+		AddField("", "", "").
+		AddField("lower", "", "").
+		AddField("Good", nil, "").
+		Build()
+	if err == nil {
+		t.Fatal("expected errors")
+	}
+	msg := err.Error()
+	if !strings.Contains(msg, "must not be empty") {
+		t.Errorf("expected empty name error in: %v", msg)
+	}
+	if !strings.Contains(msg, "must be exported") {
+		t.Errorf("expected unexported error in: %v", msg)
+	}
+	if !strings.Contains(msg, "must not be nil") {
+		t.Errorf("expected nil type error in: %v", msg)
+	}
+}
+
+func TestBuilder_Build_DeterministicOrder(t *testing.T) {
+	// Build multiple times and verify the field order is always the same.
+	for i := 0; i < 100; i++ {
+		ds, err := NewBuilder().
+			AddField("Zebra", "", "").
+			AddField("Alpha", 0, "").
+			AddField("Middle", false, "").
+			Build()
+		if err != nil {
+			t.Fatalf("Build() unexpected error: %v", err)
+		}
+
+		typ := ds.definition
+		names := make([]string, typ.NumField())
+		for j := 0; j < typ.NumField(); j++ {
+			names[j] = typ.Field(j).Name
+		}
+		expected := []string{"Alpha", "Middle", "Zebra"}
+		if !reflect.DeepEqual(names, expected) {
+			t.Fatalf("iteration %d: expected field order %v, got %v", i, expected, names)
+		}
 	}
 }
